pyproject.toml
[tool.poetry]
name = "feature-discovery-agent"
version = "0.1.0"
description = "An AI-powered assistant that helps users discover and learn about SaaS features"
authors = ["Arya Suneesh <aryasuneesh3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = ">=0.95.0"
uvicorn = {extras = ["standard"], version = ">=0.20.0"}
pydantic = {extras = ["email"], version = ">=2.0.0"}
sqlalchemy = ">=2.0.0"
python-dotenv = ">=1.0.0"
psycopg2-binary = ">=2.9.5"
langchain = ">=0.0.310"
openai = ">=1.0.0"
tiktoken = ">=0.4.0"
beautifulsoup4 = ">=4.11.0"
lxml = ">=4.9.0"
requests = ">=2.28.0"
langchain-openai = "^0.3.8"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.3.1"
httpx = ">=0.24.0"
black = ">=23.3.0"
isort = ">=5.12.0"
flake8 = ">=6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"